refactor: move LoadingContext and LoadingProvider to shared/contexts
