+++ src/index.app.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import App from './App'
 import './i18n'
-import { LoadingProvider } from './sections/loading/LoadingProvider'
+import { LoadingProvider } from './shared/contexts/loading/LoadingProvider'
 import { ThemeProvider } from '@iqss/dataverse-design-system'
 import { AppLoader } from './sections/shared/layout/app-loader/AppLoader'
