+++ src/sections/dataset/Dataset.tsx
@@ -3,7 +3,7 @@ import { Col, Row, Tabs } from '@iqss/dataverse-design-system'
 import styles from './Dataset.module.scss'
 import { useNavigate, useSearchParams } from 'react-router-dom'
 import { DatasetLabels } from './dataset-labels/DatasetLabels'
-import { useLoading } from '../loading/LoadingContext'
+import { useLoading } from '../../shared/contexts/loading/LoadingContext'
 import { DatasetSkeleton, TabsSkeleton } from './DatasetSkeleton'
 import { NotFoundPage } from '../not-found-page/NotFoundPage'
 import { useTranslation } from 'react-i18next'
