+++ src/sections/edit-collection-featured-items/EditFeaturedItems.tsx
@@ -4,7 +4,7 @@ import { Alert } from '@iqss/dataverse-design-system'
 import { CollectionRepository } from '@/collection/domain/repositories/CollectionRepository'
 import { useGetFeaturedItems } from '../collection/useGetFeaturedItems'
 import { useCollection } from '../collection/useCollection'
-import { useLoading } from '../loading/LoadingContext'
+import { useLoading } from '../../shared/contexts/loading/LoadingContext'
 import { BreadcrumbsGenerator } from '../shared/hierarchy/BreadcrumbsGenerator'
 import { SeparationLine } from '../shared/layout/SeparationLine/SeparationLine'
 import { NotFoundPage } from '../not-found-page/NotFoundPage'
