+++ src/sections/edit-collection/EditCollection.tsx
@@ -4,7 +4,7 @@ import { Alert } from '@iqss/dataverse-design-system'
 import { CollectionRepository } from '@/collection/domain/repositories/CollectionRepository'
 import { MetadataBlockInfoRepository } from '@/metadata-block-info/domain/repositories/MetadataBlockInfoRepository'
 import { useGetCollectionUserPermissions } from '@/shared/hooks/useGetCollectionUserPermissions'
-import { useLoading } from '../loading/LoadingContext'
+import { useLoading } from '../../shared/contexts/loading/LoadingContext'
 import { useSession } from '../session/SessionContext'
 import { useCollection } from '../collection/useCollection'
 import { User } from '@/users/domain/models/User'
